Add files via upload
I have done this analysis on IPL dataset from a historical point of view, giving an overview of what has happened in the IPL over the years. I have used tools such as Pandas, Matplotlib along with Python to give a visual as well as numeric representation of the data.